write more unit tests and rearrange tests a bit
